server/requirements.txt
flask
Flask-Cors
python-docx
requests
mysqlclient
mammoth
nltk
sentence-transformers
python-dotenv
psycopg2